+++ llm_assistant/tests/__init__.py
@@ -1,3 +1,4 @@
 from . import test_prompt_arguments
 from . import test_loop_control
 from . import test_pause
+from . import test_llm_retry